feat: Overlay interfaces, component visibility tracking, p_finduid nesting no-op
