+++ data/src/scripts/_test/scripts/cheats/debug_teles.rs2
@@ -1,0 +1,2 @@
+[debugproc,hillgiant]
+p_telejump(0_48_153_38_38);
