+++ data/src/scripts/interface_bank/configs/bank.inv
@@ -2,4 +2,3 @@
 scope=perm
 size=^bank_total_slots
 stackall=yes
-protect=no
