+++ data/src/scripts/interface_controls/interfaces/player_controls.if
@@ -1,3 +1,5 @@
+type=overlay
+
 [com_0]
 type=text
 x=113
