+++ data/src/scripts/interface_options/interfaces/game_options.if
@@ -1,3 +1,5 @@
+type=overlay
+
 [com_0]
 type=graphic
 x=0
