+++ data/src/scripts/music/interfaces/musicplayer_ld.if
@@ -1,3 +1,5 @@
+type=overlay
+
 [com_0]
 type=text
 x=8
