+++ data/src/scripts/player/interfaces/ignore.if
@@ -1,3 +1,5 @@
+type=overlay
+
 [com_0]
 type=layer
 x=3
