+++ data/src/scripts/player/interfaces/quest_journal.if
@@ -1,3 +1,5 @@
+type=overlay
+
 [com_0]
 type=layer
 x=5
