+++ data/src/scripts/skill_combat/interfaces/magic/attack_magic_staff.if
@@ -1,3 +1,5 @@
+type=overlay
+
 [com_0]
 type=model
 x=135
