+++ data/src/scripts/skill_combat/interfaces/melee/attack_stabbing.if
@@ -1,3 +1,5 @@
+type=overlay
+
 [com_0]
 type=model
 x=136
