+++ data/src/scripts/skill_combat/interfaces/ranged/attack_bow.if
@@ -1,3 +1,5 @@
+type=overlay
+
 [com_0]
 type=model
 x=135
