+++ data/src/scripts/skill_magic/interfaces/magic.if
@@ -1,3 +1,5 @@
+type=overlay
+
 [wind_strike]
 type=graphic
 x=1
