+++ src/lostcity/entity/Player.ts
@@ -448,6 +448,7 @@ export default class Player extends PathingEntity {
     refreshModalClose = false;
     refreshModal = false;
     modalSticky = -1;
+    overlaySide: number[] = new Array(14).fill(-1);
     receivedFirstClose = false; // workaround to not close welcome screen on login
 
     interacted: boolean = false;
@@ -778,52 +779,52 @@ export default class Player extends PathingEntity {
                 this.lastInt = input;
                 this.executeScript(this.activeScript, true);
             } else if (opcode === ClientProt.IF_BUTTON) {
-                const com = data.g2();
+                const comId = data.g2();
 
-                // TODO: verify component is opened
-                const ifType = IfType.get(com);
-                if (!ifType) {
+                const com = IfType.get(comId);
+                if (typeof com === 'undefined' || !this.isComponentVisible(com)) {
                     continue;
                 }
 
-                this.lastCom = com;
+                this.lastCom = comId;
 
-                // todo: conditionally give protected access
                 if (this.resumeButtons.indexOf(this.lastCom) !== -1) {
                     if (this.activeScript) {
                         this.executeScript(this.activeScript, true);
                     }
                 } else {
-                    const script = ScriptProvider.getByTriggerSpecific(ServerTriggerType.IF_BUTTON, ifType.id, -1);
+                    const root = IfType.get(com.rootLayer);
+
+                    const script = ScriptProvider.getByTriggerSpecific(ServerTriggerType.IF_BUTTON, comId, -1);
                     if (script) {
-                        this.executeScript(ScriptRunner.init(script, this), true);
+                        this.executeScript(ScriptRunner.init(script, this), root.overlay == false);
                     } else {
                         if (Environment.LOCAL_DEV) {
-                            this.messageGame(`No trigger for [if_button,${ifType.comName}]`);
+                            this.messageGame(`No trigger for [if_button,${com.comName}]`);
                         }
                     }
                 }
             } else if (opcode === ClientProt.INV_BUTTON1 || opcode === ClientProt.INV_BUTTON2 || opcode === ClientProt.INV_BUTTON3 || opcode === ClientProt.INV_BUTTON4 || opcode === ClientProt.INV_BUTTON5) {
+                // jagex has if_button1-5
                 const item = data.g2();
                 const slot = data.g2();
-                const com = data.g2();
+                const comId = data.g2();
 
-                const ifType = IfType.get(com);
-                if (!ifType || !ifType.inventoryOptions || !ifType.inventoryOptions.length) {
+                const com = IfType.get(comId);
+                if (typeof com === 'undefined' || !com.inventoryOptions || !com.inventoryOptions.length || !this.isComponentVisible(com)) {
                     continue;
                 }
 
-                // todo: validate against inv component properties
-                if (opcode === ClientProt.INV_BUTTON1 && !ifType.inventoryOptions[0] ||
-                    opcode === ClientProt.INV_BUTTON2 && !ifType.inventoryOptions[1] ||
-                    opcode === ClientProt.INV_BUTTON3 && !ifType.inventoryOptions[2] ||
-                    opcode === ClientProt.INV_BUTTON4 && !ifType.inventoryOptions[3] ||
-                    opcode === ClientProt.INV_BUTTON5 && !ifType.inventoryOptions[4]
+                if (opcode === ClientProt.INV_BUTTON1 && !com.inventoryOptions[0] ||
+                    opcode === ClientProt.INV_BUTTON2 && !com.inventoryOptions[1] ||
+                    opcode === ClientProt.INV_BUTTON3 && !com.inventoryOptions[2] ||
+                    opcode === ClientProt.INV_BUTTON4 && !com.inventoryOptions[3] ||
+                    opcode === ClientProt.INV_BUTTON5 && !com.inventoryOptions[4]
                 ) {
                     continue;
                 }
 
-                const listener = this.invListeners.find(l => l.com === com);
+                const listener = this.invListeners.find(l => l.com === comId);
                 if (!listener) {
                     continue;
                 }
@@ -853,26 +854,28 @@ export default class Player extends PathingEntity {
                     trigger = ServerTriggerType.INV_BUTTON5;
                 }
 
-                // todo: give protected access as needed
-                const script = ScriptProvider.getByTrigger(trigger, ifType.id, -1);
+                const script = ScriptProvider.getByTrigger(trigger, comId, -1);
                 if (script) {
-                    this.executeScript(ScriptRunner.init(script, this), true);
+                    const root = IfType.get(com.rootLayer);
+
+                    this.executeScript(ScriptRunner.init(script, this), root.overlay == false);
                 } else {
                     if (Environment.LOCAL_DEV) {
-                        this.messageGame(`No trigger for [${ServerTriggerType.toString(trigger)},${ifType.comName}]`);
+                        this.messageGame(`No trigger for [${ServerTriggerType.toString(trigger)},${com.comName}]`);
                     }
                 }
             } else if (opcode === ClientProt.INV_BUTTOND) {
-                const com = data.g2();
+                // jagex has if_buttond
+                const comId = data.g2();
                 const slot = data.g2();
                 const targetSlot = data.g2();
 
-                const ifType = IfType.get(com);
-                if (!ifType) {
+                const com = IfType.get(comId);
+                if (typeof com === 'undefined' || !this.isComponentVisible(com)) {
                     continue;
                 }
 
-                const listener = this.invListeners.find(l => l.com === com);
+                const listener = this.invListeners.find(l => l.com === comId);
                 if (!listener) {
                     continue;
                 }
@@ -884,28 +887,30 @@ export default class Player extends PathingEntity {
 
                 if (this.delayed()) {
                     // do nothing; revert the client visual
-                    this.updateInvPartial(com, inv, [slot, targetSlot]);
+                    this.updateInvPartial(comId, inv, [slot, targetSlot]);
                     continue;
                 }
 
                 this.lastSlot = slot;
                 this.lastTargetSlot = targetSlot;
 
-                // todo: give protected access as needed
-                const dragTrigger = ScriptProvider.getByTrigger(ServerTriggerType.INV_BUTTOND, ifType.id);
+                const dragTrigger = ScriptProvider.getByTrigger(ServerTriggerType.INV_BUTTOND, comId);
                 if (dragTrigger) {
-                    this.executeScript(ScriptRunner.init(dragTrigger, this), true);
+                    const root = IfType.get(com.rootLayer);
+
+                    this.executeScript(ScriptRunner.init(dragTrigger, this), root.overlay == false);
                 } else {
                     if (Environment.LOCAL_DEV) {
-                        this.messageGame(`No trigger for [inv_buttond,${ifType.comName}]`);
+                        this.messageGame(`No trigger for [inv_buttond,${com.comName}]`);
                     }
                 }
             } else if (opcode === ClientProt.OPHELD1 || opcode === ClientProt.OPHELD2 || opcode === ClientProt.OPHELD3 || opcode === ClientProt.OPHELD4 || opcode === ClientProt.OPHELD5) {
                 const item = data.g2();
                 const slot = data.g2();
-                const com = data.g2();
+                const comId = data.g2();
 
-                if (!IfType.get(com)) {
+                const com = IfType.get(comId);
+                if (typeof com === 'undefined' || !this.isComponentVisible(com)) {
                     continue;
                 }
 
@@ -918,7 +923,7 @@ export default class Player extends PathingEntity {
                     continue;
                 }
 
-                const listener = this.invListeners.find(l => l.com === com);
+                const listener = this.invListeners.find(l => l.com === comId);
                 if (!listener) {
                     continue;
                 }
@@ -962,21 +967,23 @@ export default class Player extends PathingEntity {
             } else if (opcode === ClientProt.OPHELDU) {
                 const item = data.g2();
                 const slot = data.g2();
-                const com = data.g2();
+                const comId = data.g2();
                 const useItem = data.g2();
                 const useSlot = data.g2();
-                const useCom = data.g2();
+                const useComId = data.g2();
 
-                if (!IfType.get(com)) {
+                const com = IfType.get(comId);
+                if (typeof com === 'undefined' || !this.isComponentVisible(com)) {
                     continue;
                 }
 
-                if (!IfType.get(useCom)) {
+                const useCom = IfType.get(comId);
+                if (typeof useCom === 'undefined' || !this.isComponentVisible(useCom)) {
                     continue;
                 }
 
                 {
-                    const listener = this.invListeners.find(l => l.com === com);
+                    const listener = this.invListeners.find(l => l.com === comId);
                     if (!listener) {
                         continue;
                     }
@@ -988,7 +995,7 @@ export default class Player extends PathingEntity {
                 }
 
                 {
-                    const listener = this.invListeners.find(l => l.com === useCom);
+                    const listener = this.invListeners.find(l => l.com === useComId);
                     if (!listener) {
                         continue;
                     }
@@ -1051,18 +1058,20 @@ export default class Player extends PathingEntity {
             } else if (opcode === ClientProt.OPHELDT) {
                 const item = data.g2();
                 const slot = data.g2();
-                const com = data.g2();
-                const spellCom = data.g2();
+                const comId = data.g2();
+                const spellComId = data.g2();
 
-                if (!IfType.get(com)) {
+                const com = IfType.get(comId);
+                if (typeof com === 'undefined' || !this.isComponentVisible(com)) {
                     continue;
                 }
 
-                if (!IfType.get(spellCom)) {
+                const spellCom = IfType.get(comId);
+                if (typeof spellCom === 'undefined' || !this.isComponentVisible(spellCom)) {
                     continue;
                 }
 
-                const listener = this.invListeners.find(l => l.com === com);
+                const listener = this.invListeners.find(l => l.com === comId);
                 if (!listener) {
                     continue;
                 }
@@ -1082,13 +1091,12 @@ export default class Player extends PathingEntity {
                 this.clearInteraction();
                 this.closeModal();
 
-                const type = IfType.get(spellCom);
-                const script = ScriptProvider.getByTrigger(ServerTriggerType.OPHELDT, type.id, -1);
+                const script = ScriptProvider.getByTrigger(ServerTriggerType.OPHELDT, spellComId, -1);
                 if (script) {
                     this.executeScript(ScriptRunner.init(script, this), true);
                 } else {
                     if (Environment.LOCAL_DEV) {
-                        this.messageGame(`No trigger for [opheldt,${type.comName}]`);
+                        this.messageGame(`No trigger for [opheldt,${spellCom.comName}]`);
                     }
 
                     // todo: is this appropriate?
@@ -1146,9 +1154,10 @@ export default class Player extends PathingEntity {
 
                 const item = data.g2();
                 const slot = data.g2();
-                const com = data.g2();
+                const comId = data.g2();
 
-                if (!IfType.get(com)) {
+                const com = IfType.get(comId);
+                if (typeof com === 'undefined' || !this.isComponentVisible(com)) {
                     continue;
                 }
 
@@ -1160,7 +1169,7 @@ export default class Player extends PathingEntity {
                     continue;
                 }
 
-                const listener = this.invListeners.find(l => l.com === com);
+                const listener = this.invListeners.find(l => l.com === comId);
                 if (!listener) {
                     continue;
                 }
@@ -1193,9 +1202,10 @@ export default class Player extends PathingEntity {
                 const x = data.g2();
                 const z = data.g2();
                 const locId = data.g2();
-                const spellCom = data.g2();
+                const spellComId = data.g2();
 
-                if (!IfType.get(spellCom)) {
+                const spellCom = IfType.get(spellComId);
+                if (typeof spellCom === 'undefined' || !this.isComponentVisible(spellCom)) {
                     continue;
                 }
 
@@ -1219,7 +1229,7 @@ export default class Player extends PathingEntity {
                 this.clearInteraction();
                 this.closeModal();
 
-                this.setInteraction(loc, ServerTriggerType.APLOCT, spellCom);
+                this.setInteraction(loc, ServerTriggerType.APLOCT, spellComId);
                 pathfindX = loc.x;
                 pathfindZ = loc.z;
                 pathfindRequest = true;
@@ -1266,13 +1276,14 @@ export default class Player extends PathingEntity {
                 const nid = data.g2();
                 const item = data.g2();
                 const slot = data.g2();
-                const com = data.g2();
+                const comId = data.g2();
 
-                if (!IfType.get(com)) {
+                const com = IfType.get(comId);
+                if (typeof com === 'undefined' || !this.isComponentVisible(com)) {
                     continue;
                 }
 
-                const listener = this.invListeners.find(l => l.com === com);
+                const listener = this.invListeners.find(l => l.com === comId);
                 if (!listener) {
                     continue;
                 }
@@ -1307,9 +1318,10 @@ export default class Player extends PathingEntity {
                 pathfindRequest = true;
             } else if (opcode === ClientProt.OPNPCT) {
                 const nid = data.g2();
-                const spellCom = data.g2();
+                const spellComId = data.g2();
 
-                if (!IfType.get(spellCom)) {
+                const spellCom = IfType.get(spellComId);
+                if (typeof spellCom === 'undefined' || !this.isComponentVisible(spellCom)) {
                     continue;
                 }
 
@@ -1329,7 +1341,7 @@ export default class Player extends PathingEntity {
                 this.clearInteraction();
                 this.closeModal();
 
-                this.setInteraction(npc, ServerTriggerType.APNPCT, spellCom);
+                this.setInteraction(npc, ServerTriggerType.APNPCT, spellComId);
                 pathfindX = npc.x;
                 pathfindZ = npc.z;
                 pathfindRequest = true;
@@ -1383,9 +1395,10 @@ export default class Player extends PathingEntity {
 
                 const item = data.g2();
                 const slot = data.g2();
-                const com = data.g2();
+                const comId = data.g2();
 
-                if (!IfType.get(com)) {
+                const com = IfType.get(comId);
+                if (typeof com === 'undefined' || !this.isComponentVisible(com)) {
                     continue;
                 }
 
@@ -1397,7 +1410,7 @@ export default class Player extends PathingEntity {
                     continue;
                 }
 
-                const listener = this.invListeners.find(l => l.com === com);
+                const listener = this.invListeners.find(l => l.com === comId);
                 if (!listener) {
                     continue;
                 }
@@ -1430,9 +1443,10 @@ export default class Player extends PathingEntity {
                 const x = data.g2();
                 const z = data.g2();
                 const objId = data.g2();
-                const spellCom = data.g2();
+                const spellComId = data.g2();
 
-                if (!IfType.get(spellCom)) {
+                const spellCom = IfType.get(spellComId);
+                if (typeof spellCom === 'undefined' || !this.isComponentVisible(spellCom)) {
                     continue;
                 }
 
@@ -1456,7 +1470,7 @@ export default class Player extends PathingEntity {
                 this.clearInteraction();
                 this.closeModal();
 
-                this.setInteraction(obj, ServerTriggerType.APOBJT, spellCom);
+                this.setInteraction(obj, ServerTriggerType.APOBJT, spellComId);
                 pathfindX = obj.x;
                 pathfindZ = obj.z;
                 pathfindRequest = true;
@@ -1493,13 +1507,14 @@ export default class Player extends PathingEntity {
                 const pid = data.g2();
                 const item = data.g2();
                 const slot = data.g2();
-                const com = data.g2();
+                const comId = data.g2();
 
-                if (!IfType.get(com)) {
+                const com = IfType.get(comId);
+                if (typeof com === 'undefined' || !this.isComponentVisible(com)) {
                     continue;
                 }
 
-                const listener = this.invListeners.find(l => l.com === com);
+                const listener = this.invListeners.find(l => l.com === comId);
                 if (!listener) {
                     continue;
                 }
@@ -1535,9 +1550,10 @@ export default class Player extends PathingEntity {
                 pathfindRequest = true;
             } else if (opcode === ClientProt.OPPLAYERT) {
                 const pid = data.g2();
-                const spellCom = data.g2();
+                const spellComId = data.g2();
 
-                if (!IfType.get(spellCom)) {
+                const spellCom = IfType.get(spellComId);
+                if (typeof spellCom === 'undefined' || !this.isComponentVisible(spellCom)) {
                     continue;
                 }
 
@@ -1559,7 +1575,7 @@ export default class Player extends PathingEntity {
                 this.clearInteraction();
                 this.closeModal();
 
-                this.setInteraction(player, ServerTriggerType.APPLAYERT, spellCom);
+                this.setInteraction(player, ServerTriggerType.APPLAYERT, spellComId);
                 pathfindX = player.x;
                 pathfindZ = player.z;
                 pathfindRequest = true;
@@ -3730,6 +3746,15 @@ export default class Player extends PathingEntity {
         // todo: interpolate over time? instant teleport? verify with true tile on osrs
         this.x = endX;
         this.z = endZ;
+    }
+
+    setTab(com: number, tab: number) {
+        this.overlaySide[tab] = com;
+        this.ifSetTab(com, tab);
+    }
+
+    isComponentVisible(com: IfType) {
+        return this.modalTop === com.rootLayer || this.modalBottom === com.rootLayer || this.modalSidebar === com.rootLayer || this.overlaySide.findIndex(l => l === com.rootLayer) !== -1 || this.modalSticky === com.rootLayer;
     }
 
     // ----
